fix(sentry): parse env vars as float

See merge request ngrave/rana/app/be!231

+++ apps/morty/src/main.ts
@@ -23,10 +23,10 @@ async function bootstrap() {
   app.setGlobalPrefix(globalPrefix)
   Sentry.init({
     dsn: process.env.SENTRY_DSN,
-    tracesSampleRate: parseInt(process.env.SENTRY_TRACES_SAMPLE_RATE) || 0.5,
+    tracesSampleRate: parseFloat(process.env.SENTRY_TRACES_SAMPLE_RATE) || 0.5,
     // Set sampling rate for profiling - this is relative to tracesSampleRate
     profilesSampleRate:
-      parseInt(process.env.SENTRY_PROFILES_SAMPLE_RATE) || 1.0,
+      parseFloat(process.env.SENTRY_PROFILES_SAMPLE_RATE) || 1.0,
     environment: process.env.SENTRY_ENVIRONMENT || 'dev',
     integrations: [
       // enable HTTP calls tracing
